Implement Reusable handle change in existing participant view
