+++ frontend/.babelrc
@@ -3,6 +3,6 @@
   "plugins": [
     ["@babel/plugin-transform-runtime", { "regenerator": true }],
     ["@babel/plugin-proposal-decorators", { "legacy": true }],
-    ["@babel/plugin-proposal-class-properties", { "loose": true }]
+    ["@babel/plugin-proposal-class-properties", { "loose": true }, "classes-alias"]
   ]
 }
