+++ frontend/src/stores/ParticipantStore.js
@@ -138,10 +138,23 @@ export class ParticipantStore {
     this.participant.last_four_ssn = data
   }
 
-  // Visit State Actions
-  @action handleVisitChange = ({ name, value }) => {
-    this.visit[name] = value
+  @action setVisitProgram = data => {
+    this.setVisitService("")
+    const serviceListing = this.programs.find(program => program.id === data)
+    this.visit.program = data
+    this.setServiceList(serviceListing.services)
   }
+
+  @action handleVisitChange = ({ name, value }) => {
+    switch (name) {
+      case "program":
+        this.setVisitProgram(value)
+        break
+      default:
+        this.visit[name] = value
+    }
+  }
+
   @action setVisitService = data => {
     this.visit.service = data
   }
@@ -273,6 +286,7 @@ export class ParticipantStore {
       )
       if (ok && data) {
         this.setParticipant(data)
+        this.setIsEditing(false)
       }
     } catch (error) {
       throw "ParticipantStore:  updateParticipant() Failed  =>  " + error
